Round 11 deep code audit: fix 54 issues across runtime, lowerer, and infer
Runtime (Zig) - 13 files:
- P0-1: js_string.zig split("") infinite loop fix
- P0-2: js_date.zig arithmetic overflow panic fix
- P0-3: js_collections.zig Array.from negative length panic fix
- P0-4: js_bigint.zig BigInt.pow u32 overflow panic fix
- P0-5 through P0-10: various panic and overflow fixes
- P1-1 through P1-10: runtime safety improvements
- P2-1 through P2-8: runtime edge case fixes
- P2-4: js_symbol.zig init/symbolFor now return error union

Lowerer/Emit (Rust) - 10 files:
- P0-5/P0-6/P0-7: array callback and method fixes
- P0-8: member expression type handling
- P0-9/P0-10: operator precedence and type coercion
- P1-5 through P1-13: array method and destructure fixes
- P2-14/P2-16/P2-17: optional chaining and control flow fixes

Infer (Rust) - 5 files:
- P1-B6: JSDoc number type now f64
- P1-B7: default return type now JsAny for non-export functions
- P1-B11/B12: native builtin return type fixes
- P2-B1 through B5: expression type inference fixes
- P2-B14: ZigType variant additions
- P2-B17 through B20: panic safety improvements

Showcase compilation fixes (4 bugs):
- coerce_jsany_result_type: wrap non-JsAny return expressions in JsAny.from()
  when function return type is JsAny (P1-B7 cascade)
- coerce_i64_result_type: extend to handle JsAny-returning calls with .asI64()
  when function return type is i64
- emit_one_arg: fix double JsAny.from() wrapping in array literal spread
- emit_symbol_builtin: wrap JsSymbol.init/symbolFor with catch to handle
  error union return type from P2-4

Verification: 641 Rust tests, 29 bridge tests, E2E builds pass,
showcase double-build + zig build clean

+++ examples/test-bin-project/src/host.rs
@@ -18,15 +18,15 @@ static HANDLE: Lazy<Handle> = Lazy::new(|| RUNTIME.handle().clone());
 
 // ── Sync host functions ─────────────────────────────
 
-/// Add two integers (no SDK types needed — plain C ABI).
+/// Add two numbers (no SDK types needed — plain C ABI).
 #[host_fn]
-fn host_add(a: i64, b: i64) -> i64 {
+fn host_add(a: f64, b: f64) -> f64 {
     a + b
 }
 
-/// Multiply two integers.
+/// Multiply two numbers.
 #[host_fn]
-fn host_multiply(a: i64, b: i64) -> i64 {
+fn host_multiply(a: f64, b: f64) -> f64 {
     a * b
 }
 
@@ -38,8 +38,8 @@ fn host_concat(s1: HostStr, s2: HostStr) -> JsStr {
 
 /// Return string length.
 #[host_fn]
-fn host_strlen(s: HostStr) -> i64 {
-    s.len() as i64
+fn host_strlen(s: HostStr) -> f64 {
+    s.len() as f64
 }
 
 // ── Async host functions ────────────────────────────
